+++ app/src/main/java/id/reishandy/fueltracker/ui/component/DetailStats.kt
@@ -30,7 +30,7 @@ fun DetailStats(
             year = 0,
             maxFuelCapacity = 0.0
         ),
-        latestOdometer = 0.0,
+        latestOdometer = 0,
         averageFuelEconomy = 0.0,
         totalFuelAdded = 0.0,
         totalSpent = 0.0,
@@ -53,7 +53,7 @@ fun DetailStats(
                 title = R.string.odometer,
                 value = stringResource(
                     R.string.km_abbr_value,
-                    formatNumber(vehicleWithStats.latestOdometer)
+                    formatNumber(vehicleWithStats.latestOdometer.toDouble())
                 )
             )
 
@@ -156,7 +156,7 @@ fun DetailStatsPreview() {
                         year = 2024,
                         maxFuelCapacity = 8.0
                     ),
-                    latestOdometer = 15000.0,
+                    latestOdometer = 15000,
                     averageFuelEconomy = 39.5,
                     totalFuelAdded = 120.0,
                     totalSpent = 1500000.0,
